refactor: move shared failover config to its own module
The failover helpers resolved the QuickNode URL inline in shared/constants/network.ts, so getFailoverUrlsForInfuraNetwork was called intra-module. That made the metamask-controller integration test unable to mock it via jest.spyOn, so default networks came back with no failovers.

Moving the shared config (chain to infura-name map + the two lookup helpers) into shared/constants/network-failover.ts makes that call inter-module again, so the spy works and the test passes. It also gives the shared failover config a cohesive home.
